+++ balance-vue/src/views/flowable/processDefinition/index.vue
@@ -296,12 +296,18 @@ export default {
         Message.error('只能选择一条记录')
         return
       }
-      deleteAction('/flowable/processDefinition/delete', {
-        processDefinitionId: ids.toString(),
-        cascade: cascade
-      }).then(({ msg }) => {
-        Message.success(msg)
-        this.list()
+      this.$confirm('此操作将永久删除该记录, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteAction('/flowable/processDefinition/delete', {
+          processDefinitionId: ids.toString(),
+          cascade: cascade
+        }).then(({ msg }) => {
+          Message.success(msg)
+          this.list()
+        })
       })
     },
     btnSuspendOrActivate(processDefinitionId, suspend) {
