+++ balance-vue/src/views/flowable/processInstance/index.vue
@@ -345,12 +345,18 @@ export default {
         Message.error('请选择要删除的记录')
         return
       }
-      deleteAction('/flowable/processInstance/delete', {
-        processInstanceId: id,
-        cascade: cascade
-      }).then(({ msg }) => {
-        Message.success(msg)
-        this.list()
+      this.$confirm('此操作将永久删除该记录, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteAction('/flowable/processInstance/delete', {
+          processInstanceId: id,
+          cascade: cascade
+        }).then(({ msg }) => {
+          Message.success(msg)
+          this.list()
+        })
       })
     },
     btnSuspendOrActivate(processInstanceId, suspend) {
