+++ balance-vue/src/views/sys/codeType/modules/DicInfo.vue
@@ -319,9 +319,15 @@ export default {
         Message.error('请选择要删除的记录')
         return
       }
-      deleteAction('/sys/codeInfo/delete', { ids: ids.toString() }).then(({ msg }) => {
-        Message.success(msg)
-        this.listCodeInfo()
+      this.$confirm('此操作将永久删除该记录, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteAction('/sys/codeInfo/delete', { ids: ids.toString() }).then(({ msg }) => {
+          Message.success(msg)
+          this.listCodeInfo()
+        })
       })
     },
     selectionChange(selectedRecords) {
