+++ balance-vue/src/views/sys/menu/index.vue
@@ -279,13 +279,19 @@ export default {
         Message.error('请先删除叶子节点')
         return
       }
-      deleteAction('/sys/menu/delete', { id: this.temp.menuId }).then(({ msg }) => {
-        Message.success(msg)
-        this.resetTemp()
-        this.$nextTick(() => {
-          this.$refs['dataForm'].clearValidate()
+      this.$confirm('此操作将永久删除该记录, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteAction('/sys/menu/delete', { id: this.temp.menuId }).then(({ msg }) => {
+          Message.success(msg)
+          this.resetTemp()
+          this.$nextTick(() => {
+            this.$refs['dataForm'].clearValidate()
+          })
+          this.getTreeData()
         })
-        this.getTreeData()
       })
     }
   }
